ask/pyproject.toml
[project]
name = "ask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "google-api-core>=2.24.2",
    "google-genai>=1.10.0",
]
